llm-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.28.0
openai>=1.30.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
